Merged PR 6232044: MBuf: free on drop
### Description (Describe **what** was changed and don't skimp on the **how**)

First of all, this PR completely revamps the `MBuf` struct's API, to make it more Rust-y and, most importantly - safer.
A couple of notable changes:
* Doesn't use the `raw!` macro. It seems to have been copy+pasted (like, IMHO, too much of our codebase) from here: https://github.com/flier/rust-dpdk/blob/master/rte/src/utils.rs . It doesn't really give us anything and actually creates a very unsafe API by default.
* Implements `Deref<Target = [u8]>` (and `DerefMut`) - means that an `MBuf` is generally interchangeable with `&[u8]` in a lot of cases (like `Vec<u8>`).
* **DOESN'T IMPLEMENT COPY** - this means that we now have to track the MBuf's owner, and can't just create multiple MBufs pointing to the same struct in memory
* **FREE ON DROP** - owing to the previous bullet, we can now `impl Drop` for `MBuf` and call the `rte_pktmbuf_free` there, so that the MBuf's lifetime ties into Rust's ownership model, and we won't be able to accidentally leak unused buffers.
* `Clone` is implemented by calling into `rte_pktmbuf_clone`, which creates a shallow, ref-counted clone, so the performance overhead of cloning `MBuf`s should be similar to the STD's `Arc`.

#### Perf:
Two benchmarks (`bench_empty` and `bench_syn_auth_with_trigger`) have shown 20-30% regressions in runtime. This could very well be (also) due to the fact that the benchmarks themselves had to be adapted to no longer (illegally) re-use buffers that were already transmitted. It's hard to separate the performance of the benchmark itself from the performance of the code it is benchmarking, but this PR is IMHO probably important enough for avoiding future memory leak bugs that it is acceptable to not investigate further.

### Checklist
  - [ ] Code changes are covered by Unit Tests
  - [ ] Code changes are covered by Integration Tests

Related work items: #10357334