Merged PR 6484964: Replace cycle time (rdtsc) with monotonic time
### Description (Describe **what** was changed and don't skimp on the **how**)
Using the native cycle count (rdtsc) as clock on Hyper-V is risky, as there is no guarantee the clock will be invariant/monotonic. The best solution seems to be Hyper-V's "tsc reference page" clock source which should be relatively accurate and `clock_gettime` with `CLOCK_MONOTONIC_RAW` which guarantees monotonicity.

### Checklist
  - [x] Code changes are covered by Unit Tests
  - [x] Code changes are covered by Integration Tests

Related work items: #13899468