Merged PR 6391722: Attach mempool lifetime to mbuf
### Description (Describe **what** was changed and don't skimp on the **how**)

Previously, this pseudo-code would compile:
```rust
let mbuf = {
    let mempool = MemoryPool::create();
    MBuf::new(&mempool)
};
// at this point, `mempool` has already been dropped and free, so accessing this mbuf is invalid read-after-free
dbg!(mbuf);
```

This PR makes `&MemoryPool` impl `Allocator`, instead of `MemoryPool`, which means that the mbuf is tied to the memory pool. Next, we put the `MemoryPool` created in `main()` in an `Arc`, so it can be safely passed out to the lcores, and will only free the memory pool if and when all references to it are expired.

### Checklist
  - [ ] Code changes are covered by Unit Tests
  - [ ] Code changes are covered by Integration Tests

Related work items: #10357334