Merged PR 6559854: Add mbuf usage metric
### Description (Describe **what** was changed and don't skimp on the **how**)

Add a metric for monitoring the percentage of MBufs from the pool's total capacity, that are currently allocated.

The metrics will be emitted by a periodical task that runs as part of the control core. The main purpose of the metric is to allow a simple "sanity check", to make sure that we are, on average, using a constant number of MBufs at any given moment. Since we do not need extra precision for this metric, we decided it doesn't have to be emitted after every iteration of the main network loop.

### Checklist
  - [ ] Code changes are covered by Unit Tests
  - [ ] Code changes are covered by Integration Tests

Related work items: #14823691